Divide admin.py
