+++ opac/admin/__init__.py
@@ -1,0 +1,1 @@
+import opac.admin.models
