+++ opac/admin/models/__init__.py
@@ -1,0 +1,2 @@
+import opac.admin.models.masters
+import opac.admin.models.transactions
