+++ opac/admin/models/masters/__init__.py
@@ -1,0 +1,7 @@
+import opac.admin.models.masters.author
+import opac.admin.models.masters.book
+import opac.admin.models.masters.library
+import opac.admin.models.masters.publisher
+import opac.admin.models.masters.stock
+import opac.admin.models.masters.translator
+import opac.admin.models.masters.user
