+++ opac/admin/models/transactions/__init__.py
@@ -1,0 +1,4 @@
+import opac.admin.models.transactions.holding
+import opac.admin.models.transactions.lending
+import opac.admin.models.transactions.renewing
+import opac.admin.models.transactions.reservation
